make dynamo module, move the files
