+++ modules/api/src/test/scala/vinyldns/api/domain/batch/BatchChangeInterfacesSpec.scala
@@ -20,7 +20,7 @@ import cats.implicits._
 import org.scalatest.{Matchers, WordSpec}
 import vinyldns.api.CatsHelpers
 import vinyldns.api.domain.batch.BatchChangeInterfaces._
-import vinyldns.api.repository.dynamodb.DynamoDBRetriesExhaustedException
+import vinyldns.dynamo.repository.DynamoDBRetriesExhaustedException
 
 import cats.effect._
 import cats.implicits._
