+++ modules/dynamo/src/main/scala/vinyldns/dynamo/repository/DynamoDBHelper.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package vinyldns.api.repository.dynamodb
+package vinyldns.dynamo.repository
 
 import java.util.Collections
